Pipe activity callbacks through ImageQualityActivity
