+++ app/src/main/java/edu/washington/cs/ubicomplab/rdt_reader/ImageQualityActivity.java
@@ -1,12 +1,39 @@
 package edu.washington.cs.ubicomplab.rdt_reader;
 
 import android.app.Activity;
+import android.content.Intent;
 import android.os.Bundle;
 
 public class ImageQualityActivity extends Activity {
+    ImageQualityView mImageQualityView;
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_image_quality);
+        mImageQualityView = findViewById(R.id.imageQualityView);
+    }
+
+
+    @Override
+    public void onResume() {
+        super.onResume();
+        mImageQualityView.onResume();
+    }
+
+    @Override
+    public void onPause() {
+        super.onPause();
+        mImageQualityView.onPause();
+    }
+
+    @Override
+    public void onBackPressed() {
+        if (mImageQualityView.isExternalIntent()) {
+            super.onBackPressed();
+        } else {
+            Intent intent = new Intent(this, MainActivity.class);
+            startActivity(intent);
+        }
     }
 }
